wip: add spline option, apply spline interpolation
